fix: 세팅용 함수 커스텀훅으로 통일
