+++ src/views/@common/components/Settings.tsx
@@ -1,20 +1,13 @@
-import { useEffect } from 'react';
 import { Outlet } from 'react-router-dom';
 
+import useSetGoogleAnalytics from '../hooks/useSetGoogleAnalytics';
 import useSetInterceptors from '../hooks/useSetInterceptors';
-import setGoogleAnalytics from '../utils/setGoogleAnalytics';
-import setScreenSize from '../utils/setScreenSize';
+import useSetScreenSize from '../hooks/useSetScreenSize';
 
 const Settings = () => {
   useSetInterceptors();
-  useEffect(() => {
-    setScreenSize();
-    setGoogleAnalytics();
-    window.addEventListener('resize', setScreenSize);
-    return () => {
-      window.removeEventListener('resize', setScreenSize);
-    };
-  }, []);
+  useSetScreenSize();
+  useSetGoogleAnalytics();
 
   return <Outlet />;
 };
