--- src/views/@common/utils/setScreenSize.ts
@@ -1,8 +1,0 @@
-const setScreenSize = () => {
-  // window width 관련
-  const windowWidth = window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
-  const maxWidth = Math.min(375, windowWidth);
-  document.documentElement.style.setProperty('--app-max-width', `${maxWidth}px`);
-};
-
-export default setScreenSize;
